refactor(app-router): keep announcer ownership local
Route announcements still use the approved visible-commit boundary, but the branch has only one DOM announcer consumer. A shared acquisition API and reference counting add a second lifecycle contract without consolidating the separate Pages Router work.

Keep host creation and cleanup inside AppRouterAnnouncer to match the Next.js component shape. Remove the unused shared primitive and its implementation-level ownership test.
